+++ NanoXLSX.Core/AssemblyInfo.cs
@@ -2,3 +2,4 @@
 
 [assembly: InternalsVisibleTo("NanoXLSX.Writer")]
 [assembly: InternalsVisibleTo("NanoXLSX.Reader")]
+[assembly: InternalsVisibleTo("NanoXLSX.Core.Test")]