+++ NanoXLSX.Reader/Internal/Readers/MetadataAppReader.cs
@@ -42,7 +42,7 @@ namespace NanoXLSX.Internal.Readers
 
         #region methods
         /// <summary>
-        /// Reads the XML file form the passed stream and processes the AppData section
+        /// Reads the XML file form the passed stream and processes the AppData section. The existence of the stream should be checked previously
         /// </summary>
         /// \remark <remarks>This method is virtual. Plug-in packages may override it</remarks>
         /// <param name="stream">Stream of the XML file</param>
@@ -50,11 +50,6 @@ namespace NanoXLSX.Internal.Readers
         public virtual void Read(MemoryStream stream)
         {
             PreRead(stream);
-            if (stream == null)
-            {
-                // No metadata available in xlsx file
-                return;
-            }
             try
             {
                 using (stream) // Close after processing