+++ Changelog.md
@@ -1,5 +1,32 @@
 # Change Log
 
+## v2.6.3 
+
+---
+Release Date: **XX.YY.2025**
+
+- Changed handling of reading workbooks, when docProps are missing (formal change)
+- Added test case
+
+## v2.6.2 
+
+---
+Release Date: **24.01.2025**
+
+- Fixed a regression bug in the Cell function ConvertArray
+- Added test cases
+
+## v2.6.1 
+
+---
+Release Date: **19.01.2025**
+
+- Fixed a bug on writing default column styles (not persisted in some cases)
+- Adapted style reader: When a workbook is loaded, not defined color values of Border styles are now empty strings (were null), as if a new style is created 
+- Code maintenance
+
+Note: The color values of Border styles are handled identical on writing XLSX files, either if null or empty. The change of the reader behavior was to enforce the "What You Can Write Is What You Can Read" policy of the library (writing an empty string as color value should lead to an empty string on read).
+
 ## v2.6.0 
 
 ---